feat: reports permission implemented

+++ app/Http/Controllers/Admin/ReportsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Entities\Report;
 use App\Http\Controllers\Controller;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
@@ -22,7 +23,7 @@ class ReportsController extends Controller
     {
         $this->middleware([
             'auth',
-            'role:Super|Admin'
+            'role:Admin'
         ]);
     }
 
@@ -30,9 +31,12 @@ class ReportsController extends Controller
      * Report index view.
      *
      * @return View
+     * @throws AuthorizationException
      */
     public function index(): View
     {
+        $this->authorize('viewAny', Report::class);
+
         $reports = Report::all();
 
         return view('admin.reports.index', compact('reports'));
@@ -43,19 +47,26 @@ class ReportsController extends Controller
      *
      * @param Report $report
      * @return StreamedResponse
+     * @throws AuthorizationException
      */
     public function download(Report $report): StreamedResponse
     {
+        $this->authorize('download', $report);
+
         return Storage::download($report->file_path);
     }
 
     /**
+     * Delete the specific resource.
+     *
      * @param Report $report
      * @return RedirectResponse
      * @throws Exception
      */
     public function destroy(Report $report): RedirectResponse
     {
+        $this->authorize('delete', $report);
+
         Storage::delete($report->file_path);
 
         $report->delete();
